[CREATE QUOTA] Period saving: reduce duplication
[TRELLO STORY](https://trello.com/c/Gz8poDyu/225-dit-as-a-tariff-manager-i-want-to-create-a-new-quota-and-related-data-108)
